backend/requirements.txt
aiohttp
dico-api
schedule
pyyaml
discord-webhook
databases==0.5.3
sentry-sdk
aiosqlite==0.17.0
pandas==1.3.5
pytz==2021.3
table2ascii
sanic==22.3.2
